Variants of uncertain significance (VUS) were categorized into six levels (hot, warm, tepid, cool, cold, and ice cold), according to the Association for Clinical Genomic Science (ACGS) Best Practice Guidelines. hot: 1 very strong or 1 strong + 1 supporting or 2 moderate + 1 supporting or 1 moderate + 3 supporting evidence; warm: 1 strong or 2 moderate or 1 moderate + 2 supporting or 4 supporting evidence; tepid: 1 moderate + 1 supporting or 3 supporting evidence; cool: 1 moderate or 2 supporting evidence; cold: 1 supporting evidence; ice cold: no supporting evidence (https://www.acgs.uk.com/quality/best-practice-guidelines/#VariantGuidelines). Variants classified as cool, cold, or ice cold were considered as benign-leaning VUS, unlikely to be disease-causing. Variants classified as hot, warm, or tepid were considered to be pathogenic-leaning VUS.

Variants of uncertain significance (VUS) were categorized into six levels (hot, warm, tepid, cool, cold, and ice cold), according to the Association for Clinical Genomic Science (ACGS) Best Practice Guidelines. hot: 1 very strong or 1 strong + 1 supporting or 2 moderate + 1 supporting or 1 moderate + 3 supporting evidence; warm: 1 strong or 2 moderate or 1 moderate + 2 supporting or 4 supporting evidence; tepid: 1 moderate + 1 supporting or 3 supporting evidence; cool: 1 moderate or 2 supporting evidence; cold: 1 supporting evidence; ice cold: no supporting evidence (https://www.acgs.uk.com/quality/best-practice-guidelines/#VariantGuidelines). Variants classified as cool, cold, or ice cold were considered as benign-leaning VUS, unlikely to be disease-causing. Variants classified as hot, warm, or tepid were considered to be pathogenic-leaning VUS.

Usage

VUS_classify(data, classification_col, evidence_col)

Arguments

data

DataFrame comprising fundamental variant information, evidence labeling, and classification details

classification_col

The column name for variant classification (str). Variants should be classified into five distinct categories: "P," "LP," "B," "LB," and "VUS."

evidence_col

The column name for ACMG evidence(str). The content of this column should be composed of evidence names and their strengths, connected by semicolons or comma, such as "PM2_Supporting;PM5;BP4"

Value

A new DataFrame that includes the input data and VUS classification

Examples

data("ClinGen_dataset")
ClinGen_dataset <- VUS_classify(ClinGen_dataset, "Assertion", "Applied Evidence Codes (Met)")