Automatic definition of posterior probability and odds of pathogenicity values for different strengths of evidence

Automatic definition of posterior probability and odds of pathogenicity values for different strengths of evidence

Usage

auto_select_postp(prior_probability)

Arguments

prior_probability

The prior probability of pathogenicity (proportion of P/LP variants in a set of variants)

Value

Prior_probability and OP for each evidence level

Examples

auto_select_postp(0.1)
#>   Evidence Strength Odds of pathogenicity
#> 1               PVS                   351
#> 2                PS      18.7349939951952
#> 3                PM      4.32839392791312
#> 4                PP      2.08047925438182
#> 5                BP     0.480658481882885
#> 6                BM     0.231032576205959
#> 7                BS    0.0533760512683624
#> 8               BVS   0.00284900284900285
#>   Posterior probability of pathogenicity and benignity
#> 1                                                0.975
#> 2                                     0.67550020016016
#> 3                                    0.324749849931156
#> 4                                    0.187760764369383
#> 5                                    0.949301150041276
#> 6                                    0.974972184931784
#> 7                                    0.994104293142569
#> 8                                    0.999683544303797