Establish the thresholds for each level of evidence strength

Establish the thresholds for each level of evidence strength

Usage

get_lr_threshold(data, postp_list, direction)

Arguments

data

The results of bootstrapping

postp_list

A list of posterior probability corresponding to each level of evidence strength

direction

The direction of evidence pathogenic (Pathogenic or Benign)

Value

A list of optimized thresholds

Examples

if (FALSE) { # \dontrun{
data("ClinVar_2019_dataset")
data <- add_info(ClinVar_2019_dataset, "clnsig")
local_bootstrapped_lr(data, "PrimateAI_score", "Pathogenic",0.0441, 10000, 100, 0.01, "test_dir")
postp_list <- c(0.100, 0.211, 0.608, 0.981)
lr_CI_result <- lr_CI(10000, "test_dir")
get_lr_threshold(lr_CI_result, postp_list, "Pathogenic")
} # }