The one-sided 95% confidence bound for each estimated lr+ was determined through bootstrapping iterations, enabling the assessment of evidence strength.

The one-sided 95% confidence bound for each estimated lr+ was determined through bootstrapping iterations, enabling the assessment of evidence strength.

Usage

local_bootstrapped_lr(
  input_data,
  feature,
  direction,
  alpha,
  bootstrap,
  minpoints,
  increment,
  output_dir
)

Arguments

input_data

DataFrame comprising fundamental variant information, evidence labeling, and classification details

feature

The column name that requires testing for optimizing the thresholds

direction

The direction of evidence pathogenic (Pathogenic or Benign)

alpha

Prior probability

bootstrap

The number of bootstrapping iterations

minpoints

The number of at least pathogenic and non-pathogenic variants

increment

Sliding window

output_dir

Output directory

Value

The posterior probability values for each bootstrap iteration

Examples

if (FALSE) { # \dontrun{
data("ClinVar_2019_dataset")
data <- add_info(ClinVar_2019_dataset, "clnsig")
local_bootstrapped_lr(data, "PrimateAI_score","Pathogenic", 0.0441, 10000, 100, 0.01, "test_dir")
} # }